Add back `exporter.shipped` metric
This adds the value `exporter.shipped` back to `tenzir-ctl export` based
on request from @satta. Note that the feature is deprecated and due for
removal in the future, with the `export` operator already being a
superior replacement to the `tenzir-ctl export` command.

I did not test this, as we no longer are able to easily test metrics in
this part of the code base, so I'd like to ask @satta to review this
manually.